break project into modules. Fix plotting.
