+++ flows_get_brightest/version.py
@@ -1,2 +1,2 @@
-version = "2.0.0"
-__version__ = "2.0.0"
+version = "2.0.1"
+__version__ = "2.0.1"
